add fit code to EpitaxyInterface using common LinearFitFunctions mixin class
